requirements.txt
Django==1.4.1
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.12
celery==3.0.5
dj-database-url==0.2.1
django-celery==3.0.4
django-extensions==0.9
django-social-auth==0.7.2
django-tastypie==0.9.12-alpha
gunicorn==0.14.6
httplib2==0.7.4
ipdb==0.7
ipython==0.13
kombu==2.3.2
mimeparse==0.1.3
newrelic==1.4.0.137
numpy==1.6.2
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
readline==6.2.2
redis==2.6.0
wsgiref==0.1.2
